Added structs for provisioning profiles and apps
New folder structure
Moved everything into its own module
Added Provisioning Profile and app output
